chore: Update Microsoft Graph connection and mappers
This commit updates the Microsoft Graph connection and mappers in the code. It includes changes to the `connection.js` file to add a console log for the token, updates the `getAllUsers.js` file to use the `mapMicrosoftUserToScim` function instead of `mapMicrosoftToScim`, and updates the `createUser.js` and `updateUser.js` files to use the `mapScimUserToMicrosoft` and `mapMicrosoftUserToScim` functions respectively.

The changes also include the addition of a new file `getAllGroups.js` which imports the `mapMicrosoftGroupToScim` function and exports the `getAllGroups` function.

Lastly, the commit adds a new file `egressHandler.js` in the `scim` directory, which includes an egress handler function for filtering, sorting, and pagination of data.

```